Added event-driven Say support to Ask.
In the JSON, Ask uses a version of Say that includes an 'event' key that
is not part of the normal Say class. To accomodate, created a subclass of
Say called EventSay. To prevent devs from tripping over this special
version of Say, put it in a new namespace called Specialized, and in a
new file called TropoSpecialized. Modified Ask to take either a single
EventSay or an IEnumerable<EventSay>... I kept the single say in to
avoid breaking the API, although in retrospect it probably isn't in wide
use since Say didn't work.